Update str_glue to take ... arg
str_glue() now takes multiple character vectors as well as named
arguments. Named arguments are stored as variables in an a new
environment along with the input the environment (which by default is
the parent.frame() i.e. where str_glue was called from). Add str_glue()
to the NAMESPACE. Update str_glue() roxygen2 docstring. Ran in to issues
with including braces inside a quote in what was intended to be
commented out R code, in the examples, represent the result.